initializer is now split from core-module class for better testability; adjusting tests
